updated contract

+++ smart_contract/contracts/Transactions.sol
@@ -1,22 +1,18 @@
-//SPDX-License-Identifier: Unlicense
+// SPDX-License-Identifier: UNLICENSED
+
 pragma solidity ^0.8.0;
+
+import "hardhat/console.sol";
 
 contract Transactions {
     uint256 transactionCount;
 
-    event Transfer(
-        address from,
-        address receiver,
-        uint256 amount,
-        string message,
-        uint256 timestamp,
-        string keyword
-    );
-
+    event Transfer(address from, address receiver, uint amount, string message, uint256 timestamp, string keyword);
+  
     struct TransferStruct {
         address sender;
         address receiver;
-        uint256 amount;
+        uint amount;
         string message;
         uint256 timestamp;
         string keyword;
@@ -24,14 +20,14 @@ contract Transactions {
 
     TransferStruct[] transactions;
 
-    function addToBlockchain(address payable receiver, uint256 amount, string memory message, string memory keyword) public {
+    function addToBlockchain(address payable receiver, uint amount, string memory message, string memory keyword) public {
         transactionCount += 1;
         transactions.push(TransferStruct(msg.sender, receiver, amount, message, block.timestamp, keyword));
 
         emit Transfer(msg.sender, receiver, amount, message, block.timestamp, keyword);
     }
 
-    function getAllTransactions() public view returns(TransferStruct[] memory){
+    function getAllTransactions() public view returns (TransferStruct[] memory) {
         return transactions;
     }
